requirements.txt
asgiref==3.3.4
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.6
coverage==5.5
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.4
django-admin-interface==0.16.3
django-admin-rangefilter==0.8.1
django-autoslug==1.9.8
django-colorfield==0.4.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-phonenumber-field==5.2.0
django-weasyprint==1.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
html5lib==1.1
phonenumbers==8.12.24
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
pyphen==0.11.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
uWSGI==2.0.19.1
WeasyPrint==52.5
webencodings==0.5.1
